backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pytesseract==0.3.13
Pillow==10.4.0
PyMuPDF==1.24.10
transformers==4.44.2
torch
sentencepiece==0.2.0
sacremoses==0.1.1
pydantic==2.9.1
aiofiles==24.1.0
jinja2==3.1.4
python-dotenv==1.0.1
